rename SFTTrainerConfig -> SFTConfig, sft_trainer.py -> sft.py
